refactor: Project 내부 isExpired 생성 및 파일 배치 수정 #5
